move all export types to one file
NOTE: game-list and gep-supported-games stay separated  and should not be used with TS compiler option `'isolatedModules': true`
`
