+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "@overwolf/ow-electron-packages-types",
-    "version": "0.1.0-beta.4",
+    "version": "0.1.0-beta.5",
     "description": "Type definition file for autocompletion and documentation purposes for ow-electron packages",
     "license": "MIT",
     "types": "types.d.ts",
@@ -21,7 +21,15 @@
         "url": "https://github.com/overwolf/ow-electron-packages-types"
     },
     "exports": {
-        ".": "./types.d.ts"
+        ".": {
+            "types": "./types.d.ts"
+        },
+        "./game-list": {
+            "types": "./game-list.d.ts"
+        },
+        "./gep-supported-games": {
+            "types": "./gep-supported-games.d.ts"
+        }
     },
     "publishConfig": {
         "tag": "beta"
