--- recorder/index.ts
@@ -1,2 +1,0 @@
-export * from './recording-api.interface'
-export * from './recording-api.types'
